Finalizing starting structure of testthat
Testhat is barely functional now.

Also, we should change the way we are testing that we are structuring the documentin a backwards compatible way, to something that allows us to extend it. (I.e., relying on the structure and the section and command atributes, but not everything else being equal too)